HRTECH-4346_gettingRequestStatusIdWhenRequestCreating

See merge request SharePoint/bpmflow!20

+++ BPMFlow.Infrastructure/Data/Repositories/ObjectRequestRepository.cs
@@ -112,7 +112,7 @@ public class ObjectRequestRepository : IObjectRequestRepository
         {
             Code = ++maxCode,
             RequestId = objectRequestDto.RequestId,
-            RequestStatusId = objectRequestDto.RequestStatusId,
+            RequestStatusId = (int)objectRequestDto.RequestStatusId,
             ObjectId = objectRequestDto.ObjectId,
             PeriodId = objectRequestDto.PeriodId,
             DateStart = DateTime.Now,
